bot_gen test works except extractor.extract
